Seeding products

+++ ReviewApi/Models/Review.cs
@@ -6,7 +6,7 @@ public class Review {
     public int ProductId { get; set; }
     public int CustomerId { get; set; }
     public string Title { get; set; }
-    public double Rating { get; set; } // rating is between 0.0 and 1.0
+    public decimal Rating { get; set; } // rating is between 0.0 and 1.0
     public string Description { get; set; }
     public DateTime AddedAt { get; set; }
