+++ SharedModels/ReviewDto.cs
@@ -5,7 +5,7 @@ namespace SharedModels {
         public int CustomerId { get; set; }
         public string Title { get; set; }
         public DateTime AddedAt { get; set; }
-        public double Rating { get; set; } // rating is between 0.0 and 1.0
+        public decimal Rating { get; set; } // rating is between 0.0 and 5.0
         public string Description { get; set; }
     }
 }
